backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
psycopg2-binary==2.9.6
django-environ==0.8.1
django-filter==23.2.0
djangorestframework-simplejwt==5.2.2
django-extensions==3.2.1
django-debug-toolbar==3.2.1
pytest==7.3.1
pytest-django==4.5.2
black==23.3.0
flake8==6.0.0
mypy==1.5.0
drf-yasg==1.21.5